manifest_utils.ts
