+++ packages/netsuite-adapter/src/client/sdf_client.ts
@@ -13,7 +13,7 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
-import { collections, decorators, objects as lowerdashObjects, promises, values, strings } from '@salto-io/lowerdash'
+import { collections, decorators, objects as lowerdashObjects, promises, values } from '@salto-io/lowerdash'
 import { Values, AccountId, Value } from '@salto-io/adapter-api'
 import { mkdirp, readDir, readFile, writeFile, rm, rename } from '@salto-io/file'
 import { logger } from '@salto-io/logging'
@@ -34,8 +34,6 @@ import shellQuote from 'shell-quote'
 import {
   APPLICATION_ID,
   FILE_CABINET_PATH_SEPARATOR,
-  SCRIPT_ID,
-  WORKFLOW,
 } from '../constants'
 import {
   DEFAULT_FETCH_ALL_TYPES_AT_ONCE, DEFAULT_COMMAND_TIMEOUT_IN_MINUTES,
@@ -53,11 +51,11 @@ import {
   isCustomTypeInfo, isFileCustomizationInfo, isFolderCustomizationInfo, isTemplateCustomTypeInfo,
   mergeTypeToInstances,
 } from './utils'
+import { fixManifest } from './manifest_utils'
 
 const { makeArray } = collections.array
 const { withLimitedConcurrency } = promises.array
-const { isDefined, lookupValue } = values
-const { matchAll } = strings
+const { isDefined } = values
 const { concatObjects } = lowerdashObjects
 const log = logger(module)
 
@@ -98,34 +96,6 @@ const ADDITIONAL_FILE_PATTERN = '.template.'
 export const MINUTE_IN_MILLISECONDS = 1000 * 60
 const SINGLE_OBJECT_RETRIES = 3
 const READ_CONCURRENCY = 100
-
-const TEXT_ATTRIBUTE = '#text'
-const REQUIRED_ATTRIBUTE = '@_required'
-const INVALID_DEPENDENCIES = ['ADVANCEDEXPENSEMANAGEMENT', 'SUBSCRIPTIONBILLING', 'WMSSYSTEM', 'BILLINGACCOUNTS']
-const REFERENCED_OBJECT_REGEX = new RegExp(`${SCRIPT_ID}=(?<${SCRIPT_ID}>[a-z0-9_]+(\\.[a-z0-9_]+)*)`, 'g')
-
-type RequiredDependency = {
-  typeName: string
-  dependency: string
-  value?: Value
-}
-
-type RequiredDependencyWithCondition = (
-  RequiredDependency & {
-  condition: 'byPath'
-  path: string[]
-}) | (
-  RequiredDependency & {
-  condition?: 'fullLookup' | 'byValue'
-})
-
-const REQUIRED_FEATURES: RequiredDependencyWithCondition[] = [
-  {
-    typeName: WORKFLOW,
-    dependency: 'EXPREPORTS',
-    value: 'STDRECORDSUBSIDIARYDEFAULTACCTCORPCARDEXP',
-  },
-]
 
 const baseExecutionPath = os.tmpdir()
 
@@ -205,87 +175,6 @@ type ObjectsChunk = {
   ids: string[]
   index: number
   total: number
-}
-
-const getRequiredFeatures = (customizationInfos: CustomizationInfo[]): string[] =>
-  REQUIRED_FEATURES.filter(
-    feature => customizationInfos
-      .some(custInfo => {
-        const { typeName, value } = feature
-        if (typeName !== custInfo.typeName) {
-          return false
-        }
-        if (_.isUndefined(value)) {
-          return true
-        }
-        switch (feature.condition) {
-          case 'byPath':
-            return _.get(custInfo.values, feature.path) === value
-          case 'byValue':
-            return lookupValue(custInfo.values, val => _.isEqual(val, value))
-          default:
-            return lookupValue(custInfo.values,
-              val => _.isEqual(val, value)
-              || (_.isString(val) && _.isString(value) && val.includes(value)))
-        }
-      })
-  ).map(({ dependency }) => dependency)
-
-const getRequiredObjects = (customizationInfos: CustomizationInfo[]): string[] =>
-  _.uniq(customizationInfos.flatMap(custInfo => {
-    const objName = custInfo.values[ATTRIBUTE_PREFIX + SCRIPT_ID]
-    const requiredObjects: string[] = []
-    lookupValue(custInfo.values, val => {
-      if (!_.isString(val)) {
-        return
-      }
-
-      requiredObjects.push(...Array.from(matchAll(val, REFERENCED_OBJECT_REGEX))
-        .map(r => r.groups)
-        .filter(isDefined)
-        .map(group => group[SCRIPT_ID])
-        .filter(scriptId => scriptId !== objName
-          && !scriptId.startsWith(`${objName}.`)))
-    })
-    return requiredObjects
-  }))
-
-const fixDependenciesObject = (dependencies: Value): void => {
-  dependencies.features = dependencies.features ?? {}
-  dependencies.features.feature = dependencies.features.feature ?? []
-  dependencies.objects = dependencies.objects ?? {}
-  dependencies.objects.object = dependencies.objects.object ?? []
-}
-
-const addRequiredDependencies = (
-  dependencies: Value,
-  customizationInfos: CustomizationInfo[]
-): void => {
-  const requiredFeatures = getRequiredFeatures(customizationInfos)
-    .map(feature => ({ [REQUIRED_ATTRIBUTE]: 'true', [TEXT_ATTRIBUTE]: feature }))
-  const requiredObjects = getRequiredObjects(customizationInfos)
-  if (requiredFeatures.length === 0 && requiredObjects.length === 0) {
-    return
-  }
-
-  const { features, objects } = dependencies
-  features.feature = [
-    // remove required features that are set to "required=false"
-    ..._.differenceBy(
-      makeArray(features.feature),
-      requiredFeatures,
-      item => item[TEXT_ATTRIBUTE]
-    ),
-    ...requiredFeatures,
-  ]
-  objects.object = _.union(objects.object, requiredObjects)
-}
-
-const cleanInvalidDependencies = (dependencies: Value): void => {
-  // This is done due to an SDF bug described in SALTO-1107.
-  // This function should be removed once the bug is fixed.
-  dependencies.features.feature = makeArray(dependencies.features.feature)
-    .filter(item => !INVALID_DEPENDENCIES.includes(item[TEXT_ATTRIBUTE]))
 }
 
 export default class SdfClient {
@@ -915,27 +804,7 @@ export default class SdfClient {
   ): Promise<void> {
     const manifestPath = osPath.join(projectPath, 'src', 'manifest.xml')
     const manifestContent = (await readFile(manifestPath)).toString()
-    const manifestXml = xmlParser.parse(manifestContent, { ignoreAttributes: false })
-
-    if (!_.isPlainObject(manifestXml.manifest?.dependencies)) {
-      log.warn('manifest.xml is missing dependencies tag')
-      return
-    }
-
-    const { dependencies } = manifestXml.manifest
-    fixDependenciesObject(dependencies)
-    cleanInvalidDependencies(dependencies)
-    addRequiredDependencies(dependencies, customizationInfos)
-
-    // eslint-disable-next-line new-cap
-    const fixedDependencies = new xmlParser.j2xParser({
-      ignoreAttributes: false,
-      format: true,
-    }).parse({ dependencies })
-    const fixedManifestContent = manifestContent.replace(
-      new RegExp('<dependencies>.*</dependencies>\\n?', 'gs'), fixedDependencies
-    )
-    await writeFile(manifestPath, fixedManifestContent)
+    await writeFile(manifestPath, fixManifest(manifestContent, customizationInfos))
   }
 
   private async runDeployCommands(
